fix(statistical): honest significant-decision prose on the p == alpha boundary
Removing the ULP nudge in #39 (letting holm_bonferroni own the rejection
decision) exposed a decision-prose defect it had been masking. On the Holm
step-down boundary a metric is rejected via `adjusted_p <= alpha`, so it can be
reported significant with `p == alpha` exactly. The significant `how_detected`
string hard-coded `p = {p} (< alpha {alpha})`, which is false when p equals
alpha; the nudged threshold used to make `p < alpha` literally hold.

Make the operator reflect reality with a three-way choice: strict `<` when
`p < alpha`, `<=` when `p == alpha`, `>` when `p > alpha`. Every real path keeps
`p <= alpha` on the significant branch (a Holm-rejected metric satisfies
`p <= its step threshold`; the other corrections enter the branch only under a
strict `p < alpha`), so `>` is unreachable in production — but the `significant`
override is public, and a caller forcing significance on `p > alpha` must still
get a true string, never a false `<=`. The bonferroni, none, and single-metric
paths always emit `<` and are byte-for-byte identical to main (proven by
executing both). The equivalent/inconclusive branches are untouched.

Also restore the `float(...)` cast on `_holm_step_thresholds`, dropped alongside
the nextafter line, so `SuiteReport.metric_alphas` is a plain `float` on both the
Holm and Bonferroni paths rather than mixing `np.float64` and `float`.

Tests: failing-first prose test at the exact p == alpha boundary; a suite-level
reproduction; a metric_alphas type assertion; and a prose-honesty assertion on
the forced-override p > alpha case. Full suite green on 3.10-3.13.

+++ src/evaltrust/audit/suite.py
@@ -260,7 +260,7 @@ def _holm_step_thresholds(pvalues, rejected, alpha: float) -> list[float]:
     thresholds = []
     for i in range(k):
         if rejected[i]:
-            thresholds.append(alpha / (k - rank[i]))
+            thresholds.append(float(alpha / (k - rank[i])))
         else:
-            thresholds.append(alpha / (k - n_rejected))
+            thresholds.append(float(alpha / (k - n_rejected)))
     return thresholds
